+++ packages/launcher/lib/browsers.ts
@@ -165,7 +165,7 @@ export function launch (
   url: string,
   debuggingPort: number,
   args: string[] = [],
-  defaultBrowserEnv = {},
+  browserEnv = {},
 ) {
   debug('launching browser %o', { browser, url })
 
@@ -181,7 +181,9 @@ export function launch (
 
   // allow setting default env vars such as MOZ_HEADLESS_WIDTH
   // but only if it's not already set by the environment
-  const env = Object.assign({}, defaultBrowserEnv, process.env)
+  const env = Object.assign({}, browserEnv, process.env)
+
+  debug('spawning browser with environment %o', { env })
 
   const proc = cp.spawn(browser.path, args, { stdio: ['ignore', 'pipe', 'pipe'], env })