+++ system-tests/__snapshots__/vite_dev_server_fresh_spec.ts.js
@@ -7,15 +7,15 @@ exports['@cypress/vite-dev-server react executes all of the tests for vite3.0.2-
   ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
   │ Cypress:    1.2.3                                                                              │
   │ Browser:    FooBrowser 88                                                                      │
-  │ Specs:      6 found (App.cy.jsx, AppCompilationError.cy.jsx, MissingReact.cy.jsx, MissingReact │
-  │             InSpec.cy.jsx, Rerendering.cy.jsx, Unmount.cy.jsx)                                 │
+  │ Specs:      7 found (App.cy.jsx, AppCompilationError.cy.jsx, Errors.cy.jsx, MissingReact.cy.js │
+  │             x, MissingReactInSpec.cy.jsx, Rerendering.cy.jsx, Unmount.cy.jsx)                  │
   │ Searched:   **/*.cy.{js,jsx,ts,tsx}                                                            │
   └────────────────────────────────────────────────────────────────────────────────────────────────┘
 
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  App.cy.jsx                                                                      (1 of 6)
+  Running:  App.cy.jsx                                                                      (1 of 7)
 
 
   ✓ renders hello world
@@ -46,7 +46,7 @@ exports['@cypress/vite-dev-server react executes all of the tests for vite3.0.2-
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  AppCompilationError.cy.jsx                                                      (2 of 6)
+  Running:  AppCompilationError.cy.jsx                                                      (2 of 7)
 
 
   1) An uncaught error was detected outside of a test
@@ -98,7 +98,101 @@ We dynamically generated a new test to display this failure.
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  MissingReact.cy.jsx                                                             (3 of 6)
+  Running:  Errors.cy.jsx                                                                   (3 of 7)
+
+
+  Errors
+    1) error on mount
+    2) sync error
+    3) async error
+    4) command failure
+
+
+  0 passing
+  4 failing
+
+  1) Errors
+       error on mount:
+     Error: The following error originated from your application code, not from Cypress.
+
+  > mount error
+
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
+      [stack trace lines]
+
+  2) Errors
+       sync error:
+     Error: The following error originated from your application code, not from Cypress.
+
+  > sync error
+
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
+      [stack trace lines]
+
+  3) Errors
+       async error:
+     Error: The following error originated from your application code, not from Cypress.
+
+  > async error
+
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
+      [stack trace lines]
+
+  4) Errors
+       command failure:
+     AssertionError: Timed out retrying after 50ms: Expected to find element: \`element-that-does-not-exist\`, but never found it.
+      [stack trace lines]
+
+
+
+
+  (Results)
+
+  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
+  │ Tests:        4                                                                                │
+  │ Passing:      0                                                                                │
+  │ Failing:      4                                                                                │
+  │ Pending:      0                                                                                │
+  │ Skipped:      0                                                                                │
+  │ Screenshots:  4                                                                                │
+  │ Video:        true                                                                             │
+  │ Duration:     X seconds                                                                        │
+  │ Spec Ran:     Errors.cy.jsx                                                                    │
+  └────────────────────────────────────────────────────────────────────────────────────────────────┘
+
+
+  (Screenshots)
+
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- error on mount (failed)     (1280x720)
+     .png                                                                                           
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- sync error (failed).png     (1280x720)
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- async error (failed).pn     (1280x720)
+     g                                                                                              
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- command failure (failed     (1280x720)
+     ).png                                                                                          
+
+
+  (Video)
+
+  -  Started processing:  Compressing to 32 CRF                                                     
+  -  Finished processing: /XXX/XXX/XXX/cypress/videos/Errors.cy.jsx.mp4                   (X second)
+
+
+────────────────────────────────────────────────────────────────────────────────────────────────────
+                                                                                                    
+  Running:  MissingReact.cy.jsx                                                             (4 of 7)
 
 
   1) is missing React
@@ -107,11 +201,15 @@ We dynamically generated a new test to display this failure.
   1 failing
 
   1) is missing React:
-     ReferenceError: The following error originated from your test code, not from Cypress.
+     ReferenceError: The following error originated from your application code, not from Cypress.
 
   > React is not defined
 
-When Cypress detects uncaught errors originating from your test code it will automatically fail the current test.
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
       [stack trace lines]
 
 
@@ -146,7 +244,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  MissingReactInSpec.cy.jsx                                                       (4 of 6)
+  Running:  MissingReactInSpec.cy.jsx                                                       (5 of 7)
 
 
   1) is missing React in this file
@@ -190,7 +288,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  Rerendering.cy.jsx                                                              (5 of 6)
+  Running:  Rerendering.cy.jsx                                                              (6 of 7)
 
 
   re-render
@@ -223,7 +321,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  Unmount.cy.jsx                                                                  (6 of 6)
+  Running:  Unmount.cy.jsx                                                                  (7 of 7)
 
 
   Comp with componentWillUnmount
@@ -269,6 +367,8 @@ When Cypress detects uncaught errors originating from your test code it will aut
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✖  AppCompilationError.cy.jsx               XX:XX        1        -        1        -        - │
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
+  │ ✖  Errors.cy.jsx                            XX:XX        4        -        4        -        - │
+  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✖  MissingReact.cy.jsx                      XX:XX        1        -        1        -        - │
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✖  MissingReactInSpec.cy.jsx                XX:XX        1        -        1        -        - │
@@ -277,7 +377,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✔  Unmount.cy.jsx                           XX:XX        3        3        -        -        - │
   └────────────────────────────────────────────────────────────────────────────────────────────────┘
-    ✖  3 of 6 failed (50%)                      XX:XX        8        5        3        -        -  
+    ✖  4 of 7 failed (57%)                      XX:XX       12        5        7        -        -  
 
 
 `
@@ -291,15 +391,15 @@ exports['@cypress/vite-dev-server react executes all of the tests for vite2.8.6-
   ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
   │ Cypress:    1.2.3                                                                              │
   │ Browser:    FooBrowser 88                                                                      │
-  │ Specs:      6 found (App.cy.jsx, AppCompilationError.cy.jsx, MissingReact.cy.jsx, MissingReact │
-  │             InSpec.cy.jsx, Rerendering.cy.jsx, Unmount.cy.jsx)                                 │
+  │ Specs:      7 found (App.cy.jsx, AppCompilationError.cy.jsx, Errors.cy.jsx, MissingReact.cy.js │
+  │             x, MissingReactInSpec.cy.jsx, Rerendering.cy.jsx, Unmount.cy.jsx)                  │
   │ Searched:   **/*.cy.{js,jsx,ts,tsx}                                                            │
   └────────────────────────────────────────────────────────────────────────────────────────────────┘
 
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  App.cy.jsx                                                                      (1 of 6)
+  Running:  App.cy.jsx                                                                      (1 of 7)
 
 
   ✓ renders hello world
@@ -330,7 +430,7 @@ exports['@cypress/vite-dev-server react executes all of the tests for vite2.8.6-
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  AppCompilationError.cy.jsx                                                      (2 of 6)
+  Running:  AppCompilationError.cy.jsx                                                      (2 of 7)
 
 
   1) An uncaught error was detected outside of a test
@@ -382,7 +482,101 @@ We dynamically generated a new test to display this failure.
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  MissingReact.cy.jsx                                                             (3 of 6)
+  Running:  Errors.cy.jsx                                                                   (3 of 7)
+
+
+  Errors
+    1) error on mount
+    2) sync error
+    3) async error
+    4) command failure
+
+
+  0 passing
+  4 failing
+
+  1) Errors
+       error on mount:
+     Error: The following error originated from your application code, not from Cypress.
+
+  > mount error
+
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
+      [stack trace lines]
+
+  2) Errors
+       sync error:
+     Error: The following error originated from your application code, not from Cypress.
+
+  > sync error
+
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
+      [stack trace lines]
+
+  3) Errors
+       async error:
+     Error: The following error originated from your application code, not from Cypress.
+
+  > async error
+
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
+      [stack trace lines]
+
+  4) Errors
+       command failure:
+     AssertionError: Timed out retrying after 50ms: Expected to find element: \`element-that-does-not-exist\`, but never found it.
+      [stack trace lines]
+
+
+
+
+  (Results)
+
+  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
+  │ Tests:        4                                                                                │
+  │ Passing:      0                                                                                │
+  │ Failing:      4                                                                                │
+  │ Pending:      0                                                                                │
+  │ Skipped:      0                                                                                │
+  │ Screenshots:  4                                                                                │
+  │ Video:        true                                                                             │
+  │ Duration:     X seconds                                                                        │
+  │ Spec Ran:     Errors.cy.jsx                                                                    │
+  └────────────────────────────────────────────────────────────────────────────────────────────────┘
+
+
+  (Screenshots)
+
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- error on mount (failed)     (1280x720)
+     .png                                                                                           
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- sync error (failed).png     (1280x720)
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- async error (failed).pn     (1280x720)
+     g                                                                                              
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- command failure (failed     (1280x720)
+     ).png                                                                                          
+
+
+  (Video)
+
+  -  Started processing:  Compressing to 32 CRF                                                     
+  -  Finished processing: /XXX/XXX/XXX/cypress/videos/Errors.cy.jsx.mp4                   (X second)
+
+
+────────────────────────────────────────────────────────────────────────────────────────────────────
+                                                                                                    
+  Running:  MissingReact.cy.jsx                                                             (4 of 7)
 
 
   1) is missing React
@@ -391,11 +585,15 @@ We dynamically generated a new test to display this failure.
   1 failing
 
   1) is missing React:
-     ReferenceError: The following error originated from your test code, not from Cypress.
+     ReferenceError: The following error originated from your application code, not from Cypress.
 
   > React is not defined
 
-When Cypress detects uncaught errors originating from your test code it will automatically fail the current test.
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
       [stack trace lines]
 
 
@@ -430,7 +628,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  MissingReactInSpec.cy.jsx                                                       (4 of 6)
+  Running:  MissingReactInSpec.cy.jsx                                                       (5 of 7)
 
 
   1) is missing React in this file
@@ -474,7 +672,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  Rerendering.cy.jsx                                                              (5 of 6)
+  Running:  Rerendering.cy.jsx                                                              (6 of 7)
 
 
   re-render
@@ -507,7 +705,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  Unmount.cy.jsx                                                                  (6 of 6)
+  Running:  Unmount.cy.jsx                                                                  (7 of 7)
 
 
   Comp with componentWillUnmount
@@ -553,6 +751,8 @@ When Cypress detects uncaught errors originating from your test code it will aut
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✖  AppCompilationError.cy.jsx               XX:XX        1        -        1        -        - │
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
+  │ ✖  Errors.cy.jsx                            XX:XX        4        -        4        -        - │
+  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✖  MissingReact.cy.jsx                      XX:XX        1        -        1        -        - │
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✖  MissingReactInSpec.cy.jsx                XX:XX        1        -        1        -        - │
@@ -561,7 +761,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✔  Unmount.cy.jsx                           XX:XX        3        3        -        -        - │
   └────────────────────────────────────────────────────────────────────────────────────────────────┘
-    ✖  3 of 6 failed (50%)                      XX:XX        8        5        3        -        -  
+    ✖  4 of 7 failed (57%)                      XX:XX       12        5        7        -        -  
 
 
 `
@@ -575,15 +775,15 @@ exports['@cypress/vite-dev-server react executes all of the tests for vite2.9.1-
   ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
   │ Cypress:    1.2.3                                                                              │
   │ Browser:    FooBrowser 88                                                                      │
-  │ Specs:      6 found (App.cy.jsx, AppCompilationError.cy.jsx, MissingReact.cy.jsx, MissingReact │
-  │             InSpec.cy.jsx, Rerendering.cy.jsx, Unmount.cy.jsx)                                 │
+  │ Specs:      7 found (App.cy.jsx, AppCompilationError.cy.jsx, Errors.cy.jsx, MissingReact.cy.js │
+  │             x, MissingReactInSpec.cy.jsx, Rerendering.cy.jsx, Unmount.cy.jsx)                  │
   │ Searched:   **/*.cy.{js,jsx,ts,tsx}                                                            │
   └────────────────────────────────────────────────────────────────────────────────────────────────┘
 
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  App.cy.jsx                                                                      (1 of 6)
+  Running:  App.cy.jsx                                                                      (1 of 7)
 
 
   ✓ renders hello world
@@ -614,7 +814,7 @@ exports['@cypress/vite-dev-server react executes all of the tests for vite2.9.1-
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  AppCompilationError.cy.jsx                                                      (2 of 6)
+  Running:  AppCompilationError.cy.jsx                                                      (2 of 7)
 
 
   1) An uncaught error was detected outside of a test
@@ -666,7 +866,101 @@ We dynamically generated a new test to display this failure.
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  MissingReact.cy.jsx                                                             (3 of 6)
+  Running:  Errors.cy.jsx                                                                   (3 of 7)
+
+
+  Errors
+    1) error on mount
+    2) sync error
+    3) async error
+    4) command failure
+
+
+  0 passing
+  4 failing
+
+  1) Errors
+       error on mount:
+     Error: The following error originated from your application code, not from Cypress.
+
+  > mount error
+
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
+      [stack trace lines]
+
+  2) Errors
+       sync error:
+     Error: The following error originated from your application code, not from Cypress.
+
+  > sync error
+
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
+      [stack trace lines]
+
+  3) Errors
+       async error:
+     Error: The following error originated from your application code, not from Cypress.
+
+  > async error
+
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
+      [stack trace lines]
+
+  4) Errors
+       command failure:
+     AssertionError: Timed out retrying after 50ms: Expected to find element: \`element-that-does-not-exist\`, but never found it.
+      [stack trace lines]
+
+
+
+
+  (Results)
+
+  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
+  │ Tests:        4                                                                                │
+  │ Passing:      0                                                                                │
+  │ Failing:      4                                                                                │
+  │ Pending:      0                                                                                │
+  │ Skipped:      0                                                                                │
+  │ Screenshots:  4                                                                                │
+  │ Video:        true                                                                             │
+  │ Duration:     X seconds                                                                        │
+  │ Spec Ran:     Errors.cy.jsx                                                                    │
+  └────────────────────────────────────────────────────────────────────────────────────────────────┘
+
+
+  (Screenshots)
+
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- error on mount (failed)     (1280x720)
+     .png                                                                                           
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- sync error (failed).png     (1280x720)
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- async error (failed).pn     (1280x720)
+     g                                                                                              
+  -  /XXX/XXX/XXX/cypress/screenshots/Errors.cy.jsx/Errors -- command failure (failed     (1280x720)
+     ).png                                                                                          
+
+
+  (Video)
+
+  -  Started processing:  Compressing to 32 CRF                                                     
+  -  Finished processing: /XXX/XXX/XXX/cypress/videos/Errors.cy.jsx.mp4                   (X second)
+
+
+────────────────────────────────────────────────────────────────────────────────────────────────────
+                                                                                                    
+  Running:  MissingReact.cy.jsx                                                             (4 of 7)
 
 
   1) is missing React
@@ -675,11 +969,15 @@ We dynamically generated a new test to display this failure.
   1 failing
 
   1) is missing React:
-     ReferenceError: The following error originated from your test code, not from Cypress.
+     ReferenceError: The following error originated from your application code, not from Cypress.
 
   > React is not defined
 
-When Cypress detects uncaught errors originating from your test code it will automatically fail the current test.
+When Cypress detects uncaught errors originating from your application it will automatically fail the current test.
+
+This behavior is configurable, and you can choose to turn this off by listening to the \`uncaught:exception\` event.
+
+https://on.cypress.io/uncaught-exception-from-application
       [stack trace lines]
 
 
@@ -714,7 +1012,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  MissingReactInSpec.cy.jsx                                                       (4 of 6)
+  Running:  MissingReactInSpec.cy.jsx                                                       (5 of 7)
 
 
   1) is missing React in this file
@@ -758,7 +1056,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  Rerendering.cy.jsx                                                              (5 of 6)
+  Running:  Rerendering.cy.jsx                                                              (6 of 7)
 
 
   re-render
@@ -791,7 +1089,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
 
 ────────────────────────────────────────────────────────────────────────────────────────────────────
                                                                                                     
-  Running:  Unmount.cy.jsx                                                                  (6 of 6)
+  Running:  Unmount.cy.jsx                                                                  (7 of 7)
 
 
   Comp with componentWillUnmount
@@ -837,6 +1135,8 @@ When Cypress detects uncaught errors originating from your test code it will aut
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✖  AppCompilationError.cy.jsx               XX:XX        1        -        1        -        - │
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
+  │ ✖  Errors.cy.jsx                            XX:XX        4        -        4        -        - │
+  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✖  MissingReact.cy.jsx                      XX:XX        1        -        1        -        - │
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✖  MissingReactInSpec.cy.jsx                XX:XX        1        -        1        -        - │
@@ -845,7 +1145,7 @@ When Cypress detects uncaught errors originating from your test code it will aut
   ├────────────────────────────────────────────────────────────────────────────────────────────────┤
   │ ✔  Unmount.cy.jsx                           XX:XX        3        3        -        -        - │
   └────────────────────────────────────────────────────────────────────────────────────────────────┘
-    ✖  3 of 6 failed (50%)                      XX:XX        8        5        3        -        -  
+    ✖  4 of 7 failed (57%)                      XX:XX       12        5        7        -        -  
 
 
 `